+++ example/lib/src/app/view/app.dart
@@ -10,13 +10,13 @@ class FlutteryExampleApp extends AppStatefulWidget {
   FlutteryExampleApp({super.key});
   // This is the 'App State object' of the application.
   @override
-  AppState createAppState() => _ExampleAppState();
+  AppStateX createAppState() => _ExampleAppState();
 }
 
 /// This is the 'View' of the application.
 /// The 'look and behavior' of the app.
 ///
-class _ExampleAppState extends AppState {
+class _ExampleAppState extends AppStateX<FlutteryExampleApp> {
   _ExampleAppState()
       : super(
           controller: ExampleAppController(),